cleanup spring config. share test profile across modules and keep production profile in webapp module
