+++ 52n-wps-commons/src/test/resources/spring/spring-db-config.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+<beans profile="test" xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xmlns:tx="http://www.springframework.org/schema/tx" xmlns:jdbc="http://www.springframework.org/schema/jdbc"
 	xsi:schemaLocation="http://www.springframework.org/schema/beans
 						http://www.springframework.org/schema/beans/spring-beans-4.1.xsd
-						http://www.springframework.org/schema/tx 
+						http://www.springframework.org/schema/tx
 						http://www.springframework.org/schema/tx/spring-tx.xsd
-						http://www.springframework.org/schema/jdbc 
+						http://www.springframework.org/schema/jdbc
 						http://www.springframework.org/schema/jdbc/spring-jdbc.xsd">
 
 	<tx:annotation-driven />
@@ -18,24 +18,7 @@
 		<constructor-arg ref="dataSource" />
 	</bean>
 
-	<beans profile="production">
-		<bean id="dataSource" class="org.apache.commons.dbcp.BasicDataSource" destroy-method="close">
-			<property name="driverClassName" value="org.hsqldb.jdbcDriver"></property>
-			<property name="url"
-				value="jdbc:hsqldb:file:${webroot}/WEB-INF/classes/db/data/wpsconfig;shutdown=true;hsqldb.write_delay=false;"></property>
-			<property name="username" value="SA"></property>
-			<property name="password" value=""></property>
-			<property name="initialSize" value="5"></property>
-			<property name="maxActive" value="10"></property>
-		</bean>
-
-		<jdbc:initialize-database data-source="dataSource" ignore-failures="ALL">
-			<jdbc:script location="classpath:db/schema.sql" />
-<!-- 			<jdbc:script location="classpath:db/initial-data.sql" /> -->
-		</jdbc:initialize-database>
-	</beans>
-
-	<beans profile="test">
+	<beans>
 		<!-- In-memory database to be used during testing -->
 		<jdbc:embedded-database id="dataSource" type="HSQL">
 			<jdbc:script location="classpath:db/schema.sql" />
