+++ 52n-wps-r/src/test/webapp/R/R_Datatype.conf
@@ -1,0 +1,15 @@
+# comma separated list for each mimetype
+# <annotation key>, <mimetype>, <format hint for input and output (e.g. file, textfile, zip)>
+rdata, application/rData, file
+rdata+Spatial, application/rData+Spatial, file
+rdata+SpatialPoints, application/rData+SpatialPoints, file
+rdata+SpatialPolygons, application/rData+SpatialPolygons, file
+text/html, text/html, file
+html, text/html, file
+json, application/json, file
+application/json, application/json, file
+csv, text/csv, file
+text/csv, text/csv, file
+text/x-markdown, text/x-markdown, file
+rnw, application/x-tex, file
+tex, application/x-tex, file
