+++ 52n-wps-webapp/src/main/resources/applicationContext.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
+<beans profile="production" xmlns="http://www.springframework.org/schema/beans"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:context="http://www.springframework.org/schema/context"
 	xmlns:mvc="http://www.springframework.org/schema/mvc"
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd 
+http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd
 http://www.springframework.org/schema/mvc http://www.springframework.org/schema/mvc/spring-mvc-3.0.xsd">
 
 	<context:annotation-config />
 
-	<import resource="spring/spring-db-config.xml" />
-	<import resource="spring/spring-security-config.xml" />
-	
-	<import resource="classpath*:**/wps-config-*.xml" />
+    <import resource="spring/spring-db-config.xml" />
+    <import resource="spring/spring-security-config.xml" />
+
+    <import resource="classpath*:**/wps-config-*.xml" />
 
 </beans>
