+++ 52n-wps-webapp/src/main/resources/dispatcher-servlet.xml
@@ -14,7 +14,7 @@ http://www.springframework.org/schema/mvc http://www.springframework.org/schema/
 
     <bean id="webProcessingService" class="org.n52.wps.server.WebProcessingService" init-method="init" />
 
-    <bean id="springIntegrationHelper" class="org.n52.wps.util.SpringIntegrationHelper" init-method="init" />
+    <bean id="springIntegrationHelper" class="org.n52.wps.commons.SpringIntegrationHelper" init-method="init" />
 
     <bean id="repositoryManager" class="org.n52.wps.server.RepositoryManager" depends-on="springIntegrationHelper" />
 
@@ -53,14 +53,6 @@ http://www.springframework.org/schema/mvc http://www.springframework.org/schema/
 		<!-- Include test modules in the context for integration testing -->
 		<context:component-scan base-package="org.n52.wps">
 			<context:include-filter type="assignable" expression="org.n52.wps.webapp.api.ConfigurationModule" />
-		</context:component-scan>
-	</beans>
-
-	<beans profile="test">
-		<!-- Include test modules in the context for integration testing -->
-		<context:component-scan base-package="org.n52.wps">
-			<context:include-filter type="assignable" expression="org.n52.wps.webapp.api.ConfigurationModule" />
-			<context:exclude-filter type="regex" expression="org\.n52\.wps\.webapp\.testmodules\..*" />
 		</context:component-scan>
 	</beans>
